+++ tiger-proxy/src/main/java/de/gematik/test/tiger/mockserver/httpclient/Http2SettingsHandler.java
@@ -16,7 +16,7 @@
 
 package de.gematik.test.tiger.mockserver.httpclient;
 
-import de.gematik.test.tiger.mockserver.model.Protocol;
+import de.gematik.test.tiger.mockserver.model.HttpProtocol;
 import io.netty.channel.ChannelHandlerContext;
 import io.netty.channel.SimpleChannelInboundHandler;
 import io.netty.handler.codec.http2.Http2Settings;
@@ -30,16 +30,16 @@ import java.util.concurrent.CompletableFuture;
 public class Http2SettingsHandler extends SimpleChannelInboundHandler<Http2Settings> {
   private final CompletableFuture<Http2Settings> settingsFuture;
 
-  public Http2SettingsHandler(CompletableFuture<Protocol> protocolFuture) {
+  public Http2SettingsHandler(CompletableFuture<HttpProtocol> protocolFuture) {
     this.settingsFuture = new CompletableFuture<>();
     settingsFuture.whenComplete(
         ((http2Settings, throwable) -> {
           if (throwable != null) {
             protocolFuture.completeExceptionally(throwable);
           } else if (http2Settings != null) {
-            protocolFuture.complete(Protocol.HTTP_2);
+            protocolFuture.complete(HttpProtocol.HTTP_2);
           } else {
-            protocolFuture.complete(Protocol.HTTP_1_1);
+            protocolFuture.complete(HttpProtocol.HTTP_1_1);
           }
         }));
   }