+++ tiger-proxy/src/main/java/de/gematik/test/tiger/mockserver/httpclient/NettyHttpClient.java
@@ -21,7 +21,7 @@ import static de.gematik.test.tiger.mockserver.model.HttpResponse.response;
 import de.gematik.test.tiger.mockserver.configuration.MockServerConfiguration;
 import de.gematik.test.tiger.mockserver.model.*;
 import de.gematik.test.tiger.mockserver.model.BinaryMessage;
-import de.gematik.test.tiger.mockserver.model.Protocol;
+import de.gematik.test.tiger.mockserver.model.HttpProtocol;
 import de.gematik.test.tiger.mockserver.proxyconfiguration.ProxyConfiguration;
 import de.gematik.test.tiger.mockserver.socket.tls.NettySslContextFactory;
 import io.netty.buffer.ByteBufUtil;
@@ -98,20 +98,20 @@ public class NettyHttpClient {
         requestInfo.setRemoteServerAddress(
             requestInfo.getDataToSend().socketAddressFromHostHeader());
       }
-      if (Protocol.HTTP_2.equals(requestInfo.getDataToSend().getProtocol())
+      if (HttpProtocol.HTTP_2.equals(requestInfo.getDataToSend().getProtocol())
           && !Boolean.TRUE.equals(requestInfo.getDataToSend().isSecure())) {
         log.warn(
             "HTTP2 requires ALPN but request is not secure (i.e. TLS) so protocol changed"
                 + " to HTTP1");
-        requestInfo.getDataToSend().setProtocol(Protocol.HTTP_1_1);
+        requestInfo.getDataToSend().setProtocol(HttpProtocol.HTTP_1_1);
       }
 
       final CompletableFuture<HttpResponse> httpResponseFuture = new CompletableFuture<>();
       final CompletableFuture<Message> responseFuture = new CompletableFuture<>();
-      final Protocol httpProtocol =
+      final HttpProtocol httpProtocol =
           requestInfo.getDataToSend().getProtocol() != null
               ? requestInfo.getDataToSend().getProtocol()
-              : Protocol.HTTP_1_1;
+              : HttpProtocol.HTTP_1_1;
 
       final HttpClientInitializer clientInitializer = createClientInitializer(httpProtocol);
 
@@ -189,7 +189,7 @@ public class NettyHttpClient {
     return sendRequest(requestInfo, configuration.socketConnectionTimeoutInMillis());
   }
 
-  public HttpClientInitializer createClientInitializer(Protocol httpProtocol) {
+  public HttpClientInitializer createClientInitializer(HttpProtocol httpProtocol) {
     return new HttpClientInitializer(
         configuration, proxyConfigurations, nettySslContextFactory, httpProtocol);
   }