+++ tiger-proxy/src/main/java/de/gematik/test/tiger/mockserver/netty/MockServer.java
@@ -64,6 +64,8 @@ public class MockServer extends LifeCycle {
   private InetSocketAddress remoteSocket;
   private HttpActionHandler actionHandler;
   private Map<SocketAddress, TigerConnectionStatus> connectionStatusMap = new ConcurrentHashMap<>();
+  private NettySslContextFactory serverSslContextFactory;
+  private NettySslContextFactory clientSslContextFactory;
 
   /**
    * Start the instance using the ports provided
@@ -205,10 +207,8 @@ public class MockServer extends LifeCycle {
       portBindings = Arrays.asList(localPorts);
     }
 
-    final NettySslContextFactory nettyServerSslContextFactory =
-        new NettySslContextFactory(configuration, true);
-    final NettySslContextFactory nettyClientSslContextFactory =
-        new NettySslContextFactory(configuration, false);
+    serverSslContextFactory = new NettySslContextFactory(configuration, true);
+    clientSslContextFactory = new NettySslContextFactory(configuration, false);
 
     actionHandler =
         new HttpActionHandler(
@@ -216,7 +216,7 @@ public class MockServer extends LifeCycle {
             getEventLoopGroup(),
             httpState,
             proxyConfigurations,
-            nettyClientSslContextFactory);
+          clientSslContextFactory);
     serverServerBootstrap =
         new ServerBootstrap()
             .group(bossGroup, workerGroup)
@@ -230,7 +230,7 @@ public class MockServer extends LifeCycle {
                 new WriteBufferWaterMark(8 * 1024, 32 * 1024))
             .childHandler(
                 new MockServerChannelInitializer(
-                    configuration, this, httpState, actionHandler, nettyServerSslContextFactory))
+                    configuration, this, httpState, actionHandler))
             .childAttr(REMOTE_SOCKET, remoteSocket)
             .childAttr(PROXYING, remoteSocket != null);
 
@@ -281,7 +281,6 @@ public class MockServer extends LifeCycle {
     private final MockServer mockServer;
     private final HttpState httpState;
     private final HttpActionHandler actionHandler;
-    private final NettySslContextFactory nettyServerSslContextFactory;
 
     @Override
     public void initChannel(SocketChannel ch) {
@@ -294,8 +293,7 @@ public class MockServer extends LifeCycle {
                   configuration,
                   mockServer,
                   httpState,
-                  actionHandler,
-                  nettyServerSslContextFactory));
+                  actionHandler));
     }
   }
 }