+++ tiger-proxy/src/main/java/de/gematik/test/tiger/mockserver/socket/tls/KeyAndCertificateFactory.java
@@ -16,24 +16,12 @@
 
 package de.gematik.test.tiger.mockserver.socket.tls;
 
-import java.security.PrivateKey;
-import java.security.cert.X509Certificate;
-import java.util.List;
+import de.gematik.test.tiger.common.pki.TigerPkiIdentity;
 
 /*
  * @author jamesdbloom
  */
 public interface KeyAndCertificateFactory {
 
-  void buildAndSavePrivateKeyAndX509Certificate();
-
-  boolean certificateNotYetCreated();
-
-  PrivateKey privateKey();
-
-  X509Certificate x509Certificate();
-
-  X509Certificate certificateAuthorityX509Certificate();
-
-  List<X509Certificate> certificateChain();
+  TigerPkiIdentity buildAndSavePrivateKeyAndX509Certificate(String hostname);
 }